pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-video-detector"
version = "0.1.0"
description = "LLM-assisted AI video detection service and CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "jinja2>=3.1.3",
  "numpy==1.26.3",
  "opencv-python>=4.9.0.80",
  "pillow>=10.2.0",
  "pydantic>=2.6.0",
  "python-multipart>=0.0.9",
  "python-dotenv>=1.0.1",
  "requests>=2.31.0",
  "uvicorn>=0.27.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
ai-video-detector = "ai_video_detector.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
